separate model and endpoint,modal dialog in endpoints
