+++ nlp-estimate-app/src/admin.js
@@ -2,14 +2,18 @@ import ForgeUI, {AdminPage, render, Text,Tabs, Tab, Button, Fragment, useState, 
 import {Notifications} from "./notifications";
 import {Configurations} from "./configurations";
 import {Endpoints} from "./endpoints";
+import {Models} from "./model";
 
 export const AdminConfigPage = () => {
   return (
     <Tabs>
       <Tab label="Bericht(e) erstellen">
       </Tab>
-      <Tab label="API Endpunkte / Modell">
+      <Tab label="API Endpunkte">
         <Endpoints/>
+      </Tab>
+      <Tab label="Modelle">
+        <Models/>
       </Tab>
       <Tab label="Benachrichtigungen">
         <Notifications/>
@@ -18,63 +22,8 @@ export const AdminConfigPage = () => {
         <Configurations/>
       </Tab>
     </Tabs>
-
   )
-
 }
-// export const AdminConfigPage = () => {
-//   const [navigation, setNavigation] = useState({
-//     report:false,
-//     endpoints:false,
-//     notifications:false,
-//     configurations:false
-//   });
-//   return (
-//     <Fragment>
-//       <Button
-//         text={`Bericht(e) erstellen`}
-//         onClick={async () => {
-//           setNavigation(Object.assign({},navigation, {
-//             report:!navigation.report
-//           }));
-//         }}
-//         icon = "page"
-//       />
-//       <Button
-//         text={`API Endpunkte / Modell`}
-//         onClick={async () => {
-//           setNavigation(Object.assign({},navigation, {
-//             endpoints:!navigation.endpoints
-//           }));
-//         }}
-//         icon = "graph-line"
-//       />
-//       {navigation.endpoints && <Endpoints/>}
-//
-//       <Button
-//         text={`Benachrichtigungen`}
-//         onClick={async () => {
-//           setNavigation(Object.assign({},navigation, {
-//             notifications:!navigation.notifications
-//           }));
-//         }}
-//         icon = "notification-all"
-//       />
-//       {navigation.notifications && <Notifications/>}
-//
-//       <Button
-//         text={`Allgemeine Konfigurationen`}
-//         onClick={async () => {
-//           setNavigation(Object.assign({},navigation, {
-//             configurations:!navigation.configurations
-//           }));
-//         }}
-//         icon = "settings"
-//       />
-//       {navigation.configurations && <Configurations/>}
-//     </Fragment>
-//   );
-// };
 
 export const run = render(
   <AdminPage>
